完善功能

+++ file_py/file_operations.py
@@ -1,3 +1,5 @@
+import os
+
 from PySide6.QtWidgets import QMessageBox
 from file_py.common import select_file
 import pandas as pd
@@ -15,44 +17,64 @@ class FileManager:
             QMessageBox.information(None, "系统消息", "请先导入数据")
 
     def add(self):
-        path = select_file("请选择要添加数据的Excel文件")
-        if path:
-            df = pd.read_excel(path, sheet_name='Sheet1', usecols=['编码', '当前版本', '历史版本'])
-            if df.empty:
-                QMessageBox.warning(None, "系统消息", "文件为空")
-                return
-            else:
-                df.rename(columns={'编码': 'Item', '当前版本': 'C_version', '历史版本': 'H_version'}, inplace=True)
-                if self.sqlHelper.inster_sqlserver(df=df) > 0:
-                    QMessageBox.information(None, "系统消息", "添加成功")
+        if QMessageBox.warning(None, "系统消息", "您点击了添加按钮，请确认是否要添加数据") == QMessageBox.Yes:
+            path = select_file("请选择要添加数据的Excel文件")
+            if path:
+                df = pd.read_excel(path, sheet_name='Sheet1', usecols=['编码', '当前版本', '历史版本'])
+                if df.empty:
+                    QMessageBox.warning(None, "系统消息", "文件为空")
+                    return
                 else:
-                    QMessageBox.warning(None, "系统消息", "添加失败")
+                    df.rename(columns={'编码': 'Item', '当前版本': 'C_version', '历史版本': 'H_version'}, inplace=True)
+                    if self.sqlHelper.inster_sqlserver(df=df) > 0:
+                        QMessageBox.information(None, "系统消息", "添加成功")
+                    else:
+                        QMessageBox.warning(None, "系统消息", "添加失败")
 
     def update(self):
-        path = select_file("请选择要添加数据的Excel文件")
-        if path:
-            df = pd.read_excel(path, sheet_name='Sheet2', usecols=['编码', '当前版本', '历史版本'])
-            if df.empty:
-                QMessageBox.warning(None, "系统消息", "文件为空")
-                return
-            else:
-                df.rename(columns={'编码': 'Item', '当前版本': 'C_version', '历史版本': 'H_version'}, inplace=True)
-                if self.sqlHelper.update_sqlserver(df=df) > 0:
-                    QMessageBox.information(None, "系统消息", "修改成功")
+        if QMessageBox.warning(None, "系统消息", "您点击了修改按钮，请确认是否要修改数据") == QMessageBox.Yes:
+            path = select_file("请选择要添加数据的Excel文件")
+            if path:
+                df = pd.read_excel(path, sheet_name='Sheet2', usecols=['编码', '当前版本', '历史版本'])
+                if df.empty:
+                    QMessageBox.warning(None, "系统消息", "文件为空")
+                    return
                 else:
-                    QMessageBox.warning(None, "系统消息", "修改失败")
+                    df.rename(columns={'编码': 'Item', '当前版本': 'C_version', '历史版本': 'H_version'}, inplace=True)
+                    if self.sqlHelper.update_sqlserver(df=df) > 0:
+                        QMessageBox.information(None, "系统消息", "修改成功")
+                    else:
+                        QMessageBox.warning(None, "系统消息", "修改失败")
 
     def delete(self):
-        path = select_file("请选择要添加数据的Excel文件")
-        if path:
-            df = pd.read_excel(path, sheet_name='Sheet3', usecols=['编码'])
+        if QMessageBox.warning(None, "系统消息", "您点击了删除按钮，请确认是否要删除数据") == QMessageBox.Yes:
+            path = select_file("请选择要添加数据的Excel文件")
+            if path:
+                df = pd.read_excel(path, sheet_name='Sheet3', usecols=['编码'])
+                if df.empty:
+                    QMessageBox.warning(None, "系统消息", "文件为空")
+                    return
+                else:
+                    df.rename(columns={'编码': 'Item'}, inplace=True)
+                    if self.sqlHelper.Delete_SQLServer(df=df) > 0:
+                        QMessageBox.information(None, "系统消息", "删除成功")
+
+    # 导出数据
+    def export(self):
+        # 询问用户是否导出数据，如果是，则导出数据
+        if QMessageBox.question(None, "系统消息", "是否导出数据") == QMessageBox.Yes:
+            df = self.query_all()
             if df.empty:
-                QMessageBox.warning(None, "系统消息", "文件为空")
-                return
+                QMessageBox.warning(None, "系统消息", "查询结果为空")
             else:
-                df.rename(columns={'编码': 'Item'}, inplace=True)
-                if self.sqlHelper.Delete_SQLServer(df=df) > 0:
-                    QMessageBox.information(None, "系统消息", "删除成功")
+                # 导出数据
+                df.to_excel('data.xlsx', index=False)
+                # 打开文件导出的文件
+                os.startfile('data.xlsx')
+
+    def query_all(self):
+        sql = "SELECT Item, C_version, H_version FROM Item_version"
+        return pd.read_sql_query(sql, self.sqliteMemory.memory_conn)
 
     def query(self, item):
         # 查询Item='02350DSD'的内存数据库中的数据
